Feat: Pturn implementation & Waypoints from txt
The positions and velocities of the pturn trajectory are generated through the create_pturn_trajectory function. The number of points in the trajectory is self.num_pturn_points. During the pturn maneuver, the number of points passed in the pturn trajectory is indicated by the value of the counter self.pturn_count.

Using argparse, you can specify a txt file using the --filename argument to read specific waypoints. The txt file consists of the number of waypoints and the x, y, z coordinates of each waypoint. The default filename is waypoint.txt, so there is no need to modify it.